REF: made TypedStream API more C++
git-svn-id: https://www.nitrc.org/svn/cmtk/trunk@3647 42a5c34f-2066-0410-bec5-ba365beb4995

+++ core/apps/gmm.cxx
@@ -235,7 +235,7 @@ doMain
       if ( ! period )
 	period = path + strlen( path );
       
-      snprintf( period, PATH_MAX - (period-path), "_prob%d%s", 1+k, outputImagePath + (period-path) );
+      snprintf( period, PATH_MAX - (period-path), "_prob%d%s", static_cast<int>( 1+k ), outputImagePath + (period-path) );
       cmtk::VolumeIO::Write( *(pMaps[k]), path );
       }
     }
